+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### 2.0.3
 
+* [Laravel4] App is rebooted before each test. Fixes #1205 *2014-07-15*
+* FIX: `codeception/specify` is now available in phar *2014-07-14*
 * FIX: Interactive console works again *2014-07-09*
 * `_bootstrap.php` is now loaded before `beforeSuite` module hooks.
 * FIX: Suite `_bootstrap.php` was loaded after test run by @samdark *2014-07-11*